Abstracted data from the main into a game object.
Added game object, which initialises the components such as shuffle, add_players to player vector and also with its own initialise method using the deck inside the game object.

So far the objective of clearing up the main has been a success, however, there must be a new print_game function to print out everything relevant to the game.

After this the next step is to create a score board of sorts, to keep track of player points/score